+++ src/services/riot_formulas.rs
@@ -42,8 +42,8 @@ pub struct RiotFormulas;
 impl RiotFormulas {
     /// Uses wiki's formula to return base stats for a given champion
     #[inline]
-    pub const fn stat_growth(base: f64, growth_per_level: f64, level: u8) -> f64 {
-        base + growth_per_level * (level as f64 - 1.0) * (0.7025 + 0.0175 * (level as f64 - 1.0))
+    pub const fn stat_growth(base: f32, growth_per_level: f32, level: u8) -> f32 {
+        base + growth_per_level * (level as f32 - 1.0) * (0.7025 + 0.0175 * (level as f32 - 1.0))
     }
     /// Percentage values are entered in this section as a number in range 0-100
     ///
@@ -55,23 +55,23 @@ impl RiotFormulas {
     /// return 0.51
     ///
     /// ```
-    pub fn percent_value<const N: usize>(from_vec: SmallVec<[f64; N]>) -> f64 {
+    pub fn percent_value<const N: usize>(from_vec: SmallVec<[f32; N]>) -> f32 {
         from_vec
             .iter()
-            .map(|value: &f64| 100.0 - value)
-            .product::<f64>()
-            / 10f64.powi((from_vec.len() << 1) as i32)
+            .map(|value: &f32| 100.0 - value)
+            .product::<f32>()
+            / 10f32.powi((from_vec.len() << 1) as i32)
     }
 
     #[inline]
-    pub const fn real_resist(percent_pen: f64, flat_pen: f64, resist: f64) -> (f64, f64) {
+    pub const fn real_resist(percent_pen: f32, flat_pen: f32, resist: f32) -> (f32, f32) {
         let real_val = (percent_pen * resist - flat_pen).max(0.0);
         let modf_val = 100.0 / (100.0 + real_val);
         (real_val, modf_val)
     }
 
     #[inline]
-    pub const fn adaptative_type(attack_damage: f64, ability_power: f64) -> AdaptativeType {
+    pub const fn adaptative_type(attack_damage: f32, ability_power: f32) -> AdaptativeType {
         if 0.35 * attack_damage >= 0.2 * ability_power {
             AdaptativeType::Physical
         } else {
